Convert timestamp at data_hdr from jiffies to ms
- Change type from __u64 to __be32

+++ linux-3.17.1/net/gmtp/mcc/loss_interval.h
@@ -30,8 +30,8 @@
 struct mcc_loss_interval {
 	__be32		 li_seqno;
 	__be32		 li_is_closed:1;
-	u32		 li_length;
-	__u64		 li_tstamp;
+	__u32		 li_length;
+	__u32		 li_tstamp;
 	__u32		 li_rtt;
 };
